chore: Changes to dash-separated filename convention
